streamlit_app/requirements_streamlit.txt
streamlit==1.32.0
streamlit-option-menu==0.3.12
scikit-learn==1.4.1.post1
tensorflow==2.15.0
keras==2.15.0
pandas==2.2.0
numpy==1.26.0
plotly==5.19.0
seaborn==0.13.2
matplotlib==3.8.3
pefile==2023.2.7
python-magic==0.4.27
python-docx==1.1.0
PyPDF2==3.0.1
reportlab==4.1.0
pillow==10.2.0
scipy==1.12.0
imbalanced-learn==0.12.0
xgboost==2.0.3
joblib==1.3.2